Refactor: move EffectiveMembership to effective_membership.rs
