+++ openraft/src/core/append_entries.rs
@@ -1,8 +1,8 @@
 use crate::core::apply_to_state_machine;
-use crate::core::EffectiveMembership;
 use crate::core::RaftCore;
 use crate::core::ServerState;
 use crate::error::AppendEntriesError;
+use crate::membership::EffectiveMembership;
 use crate::raft::AppendEntriesRequest;
 use crate::raft::AppendEntriesResponse;
 use crate::raft_types::LogIdOptionExt;
