+++ openraft/src/engine/engine.rs
@@ -4,12 +4,12 @@ use std::sync::Arc;
 
 use maplit::btreeset;
 
-use crate::core::EffectiveMembership;
 use crate::entry::RaftEntry;
 use crate::error::InitializeError;
 use crate::error::NotAMembershipEntry;
 use crate::error::NotAllowed;
 use crate::error::NotInMembers;
+use crate::membership::EffectiveMembership;
 use crate::raft_state::RaftState;
 use crate::LogId;
 use crate::LogIdOptionExt;
