+++ openraft/src/lib.rs
@@ -42,13 +42,13 @@ use serde::Serialize;
 pub use crate::config::Config;
 pub use crate::config::ConfigError;
 pub use crate::config::SnapshotPolicy;
-pub use crate::core::EffectiveMembership;
 pub use crate::core::ServerState;
 pub use crate::defensive::DefensiveCheck;
 pub use crate::defensive::DefensiveCheckBase;
 pub use crate::entry::Entry;
 pub use crate::entry::EntryPayload;
 pub use crate::entry::RaftPayload;
+pub use crate::membership::EffectiveMembership;
 pub use crate::membership::Membership;
 pub use crate::metrics::RaftMetrics;
 pub use crate::network::RPCTypes;
