+++ openraft/src/membership/mod.rs
@@ -1,3 +1,4 @@
+mod effective_membership;
 #[allow(clippy::module_inception)]
 mod membership;
 
@@ -6,5 +7,6 @@ mod membership_test;
 
 pub mod quorum;
 
+pub use effective_membership::EffectiveMembership;
 pub use membership::IntoOptionNodes;
 pub use membership::Membership;
